+++ NEWS.md
@@ -2,6 +2,7 @@
 
 - Updated vignettes to remove redundant distr6 installation via GitHub
 - Redundant `stopwarn` argument removed from `parameters()` method in `Distribution` and `ParameterSet` classes
+- Updated Kernel$new() error message to point user to Distribution class
 
 
 # distr6 1.0.0